#138 removing ThreadDatabase and link from DatabaseFactory to MediaDatabase
