+++ src/org/thoughtcrime/securesms/database/loaders/ThreadMediaLoader.java
@@ -6,8 +6,6 @@ import android.database.Cursor;
 import android.support.annotation.NonNull;
 
 import org.thoughtcrime.securesms.database.Address;
-import org.thoughtcrime.securesms.database.DatabaseFactory;
-import org.thoughtcrime.securesms.recipients.Recipient;
 import org.thoughtcrime.securesms.util.AbstractCursorLoader;
 
 public class ThreadMediaLoader extends AbstractCursorLoader {
@@ -23,10 +21,11 @@ public class ThreadMediaLoader extends AbstractCursorLoader {
 
   @Override
   public Cursor getCursor() {
-    long threadId = DatabaseFactory.getThreadDatabase(getContext()).getThreadIdFor(Recipient.from(getContext(), address, true));
-
-    if (gallery) return DatabaseFactory.getMediaDatabase(getContext()).getGalleryMediaForThread(threadId);
-    else         return DatabaseFactory.getMediaDatabase(getContext()).getDocumentMediaForThread(threadId);
+    return null;
+//    long threadId = DatabaseFactory.getThreadDatabase(getContext()).getThreadIdFor(Recipient.from(getContext(), address, true));
+//
+//    if (gallery) return DatabaseFactory.getMediaDatabase(getContext()).getGalleryMediaForThread(threadId);
+//    else         return DatabaseFactory.getMediaDatabase(getContext()).getDocumentMediaForThread(threadId);
   }
 
   public Address getAddress() {
