+++ src/org/thoughtcrime/securesms/service/DirectShareService.java
@@ -16,13 +16,10 @@ import android.support.annotation.RequiresApi;
 import android.util.Log;
 
 import org.thoughtcrime.securesms.ShareActivity;
-import org.thoughtcrime.securesms.database.DatabaseFactory;
-import org.thoughtcrime.securesms.database.ThreadDatabase;
-import org.thoughtcrime.securesms.database.model.ThreadRecord;
-import org.thoughtcrime.securesms.mms.GlideApp;
 import org.thoughtcrime.securesms.recipients.Recipient;
 import org.thoughtcrime.securesms.util.BitmapUtil;
 
+import java.util.Collections;
 import java.util.LinkedList;
 import java.util.List;
 import java.util.concurrent.ExecutionException;
@@ -36,56 +33,58 @@ public class DirectShareService extends ChooserTargetService {
   public List<ChooserTarget> onGetChooserTargets(ComponentName targetActivityName,
                                                  IntentFilter matchedFilter)
   {
-    List<ChooserTarget> results        = new LinkedList<>();
-    ComponentName       componentName  = new ComponentName(this, ShareActivity.class);
-    ThreadDatabase      threadDatabase = DatabaseFactory.getThreadDatabase(this);
-    Cursor              cursor         = threadDatabase.getDirectShareList();
-
-    try {
-      ThreadDatabase.Reader reader = threadDatabase.readerFor(cursor);
-      ThreadRecord record;
-
-      while ((record = reader.getNext()) != null && results.size() < 10) {
-          Recipient recipient = Recipient.from(this, record.getRecipient().getAddress(), false);
-          String    name      = recipient.toShortString();
-
-          Bitmap avatar;
-
-          if (recipient.getContactPhoto(getApplicationContext()) != null) {
-            try {
-              avatar = GlideApp.with(this)
-                               .asBitmap()
-                               .load(recipient.getContactPhoto(getApplicationContext()))
-                               .circleCrop()
-                               .submit(getResources().getDimensionPixelSize(android.R.dimen.notification_large_icon_width),
-                                       getResources().getDimensionPixelSize(android.R.dimen.notification_large_icon_width))
-                               .get();
-            } catch (InterruptedException | ExecutionException e) {
-              Log.w(TAG, e);
-              avatar = getFallbackDrawable(recipient);
-            }
-          } else {
-            avatar = getFallbackDrawable(recipient);
-          }
-
-          Parcel parcel = Parcel.obtain();
-          parcel.writeParcelable(recipient.getAddress(), 0);
-
-          Bundle bundle = new Bundle();
-          bundle.putLong(ShareActivity.EXTRA_THREAD_ID, record.getThreadId());
-          bundle.putByteArray(ShareActivity.EXTRA_ADDRESS_MARSHALLED, parcel.marshall());
-          bundle.putInt(ShareActivity.EXTRA_DISTRIBUTION_TYPE, record.getDistributionType());
-          bundle.setClassLoader(getClassLoader());
-
-          results.add(new ChooserTarget(name, Icon.createWithBitmap(avatar), 1.0f, componentName, bundle));
-          parcel.recycle();
-
-      }
-
-      return results;
-    } finally {
-      if (cursor != null) cursor.close();
-    }
+    // TODO: return the first 10 entries of the chat List here.
+    return Collections.EMPTY_LIST;
+//    List<ChooserTarget> results        = new LinkedList<>();
+//    ComponentName       componentName  = new ComponentName(this, ShareActivity.class);
+//    ThreadDatabase      threadDatabase = DatabaseFactory.getThreadDatabase(this);
+//    Cursor              cursor         = threadDatabase.getDirectShareList();
+//
+//    try {
+//      ThreadDatabase.Reader reader = threadDatabase.readerFor(cursor);
+//      ThreadRecord record;
+//
+//      while ((record = reader.getNext()) != null && results.size() < 10) {
+//          Recipient recipient = Recipient.from(this, record.getRecipient().getAddress(), false);
+//          String    name      = recipient.toShortString();
+//
+//          Bitmap avatar;
+//
+//          if (recipient.getContactPhoto(getApplicationContext()) != null) {
+//            try {
+//              avatar = GlideApp.with(this)
+//                               .asBitmap()
+//                               .load(recipient.getContactPhoto(getApplicationContext()))
+//                               .circleCrop()
+//                               .submit(getResources().getDimensionPixelSize(android.R.dimen.notification_large_icon_width),
+//                                       getResources().getDimensionPixelSize(android.R.dimen.notification_large_icon_width))
+//                               .get();
+//            } catch (InterruptedException | ExecutionException e) {
+//              Log.w(TAG, e);
+//              avatar = getFallbackDrawable(recipient);
+//            }
+//          } else {
+//            avatar = getFallbackDrawable(recipient);
+//          }
+//
+//          Parcel parcel = Parcel.obtain();
+//          parcel.writeParcelable(recipient.getAddress(), 0);
+//
+//          Bundle bundle = new Bundle();
+//          bundle.putLong(ShareActivity.EXTRA_THREAD_ID, record.getThreadId());
+//          bundle.putByteArray(ShareActivity.EXTRA_ADDRESS_MARSHALLED, parcel.marshall());
+//          bundle.putInt(ShareActivity.EXTRA_DISTRIBUTION_TYPE, record.getDistributionType());
+//          bundle.setClassLoader(getClassLoader());
+//
+//          results.add(new ChooserTarget(name, Icon.createWithBitmap(avatar), 1.0f, componentName, bundle));
+//          parcel.recycle();
+//
+//      }
+//
+//      return results;
+//    } finally {
+//      if (cursor != null) cursor.close();
+//    }
   }
 
   private Bitmap getFallbackDrawable(@NonNull Recipient recipient) {
